requirements.txt
Werkzeug~=3.0.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.2
scikit-learn==1.4.2
pymilvus==2.4.0
torch==2.2.2
transformers===4.39.3
pydotenv==0.0.7
pinecone-client==3.2.2
